Restore LanguagePage.jsx to feature/language-page branch
